refactor(server): inline the cloud-on-Bun bootstrap into apps/api
startBunServer was a 100-line hosted-cloud-on-Bun launcher in @epicenter/server
with exactly one caller (apps/api), and its own header said it was 'not a shared
one' (the instance composes its Bun entry directly). It was the app's composition
wearing a library coat. Inline it into apps/api/server.ts: the full bootstrap
(pool, bun:sqlite rooms, cloud auth layer, session/rooms/inference/blobs mounts,
Bun.serve) now reads top-to-bottom in the app, like worker/index.ts does, and
BunHostBindings collapses into a single app-local ApiBunBindings. The /bun barrel
loses three exports and gains mountCloudAuth + resolveRequestOAuthUser (which the
launcher used internally); each Bun entry owns its own product composition.
