+++ apps/api/server.ts
@@ -5,16 +5,24 @@
  * (`worker/index.ts`), but binds the per-concern runtime hooks to plain
  * primitives instead of Cloudflare bindings (ADR-0066):
  *
- *   - `connectDb`     a module-scope `pg.Pool` over `DATABASE_URL`
- *   - `afterResponse` fire-and-forget in the live process (no `waitUntil`)
- *   - `resolveRooms`  an in-process registry over `bun:sqlite` files
- *   - blobs           any S3 endpoint via the existing `BLOBS_S3_*` env
+ *   - the `db` leg   a module-scope `pg.Pool` over `DATABASE_URL`, drained
+ *                    fire-and-forget in the live process (no `waitUntil`)
+ *   - `resolveRooms` an in-process registry over `bun:sqlite` files
+ *   - blobs          any S3 endpoint via the existing `BLOBS_S3_*` env
  *
  * This is additive: `wrangler dev`/`deploy` still serve the Worker unchanged.
  * `bun --watch server.ts` boots instantly with real stack traces. It is the
  * hosted cloud on Bun (local dev and the runtime-parity smoke), NOT the self-host
  * artifact: the single-partition instance has its own entry
  * (`apps/self-host/server.ts`), composing no Better Auth and no Postgres (ADR-0073).
+ *
+ * The whole hosted-cloud-on-Bun bootstrap lives here, in the app, not behind a
+ * shared `@epicenter/server` factory: everything mechanical (the `pg.Pool`, the
+ * `bun:sqlite` rooms, the cloud auth layer, the session/rooms/inference/blobs
+ * mounts, `Bun.serve`) is this app's composition to own. The instance does NOT
+ * share it (it diverges on the substrate that matters: no Better Auth, no
+ * Postgres), so a shared launcher would re-introduce the mode knob ADR-0073/0074
+ * deleted. The library ships the parts; each Bun entry composes its own product.
  *
  * The wiring lives in {@link startBunApiServer} so `server.dev.ts` can boot the
  * SAME server with a dev `resolveUser` injected (the parity smoke's credential)
@@ -32,16 +40,52 @@
  * serves the dashboard in dev, and billing is the hosted Worker's concern.
  */
 
+import { mkdirSync } from 'node:fs';
+import { resolve } from 'node:path';
 import {
-	BunHostBindings,
+	bun,
+	createBunRooms,
+	createDb,
+	createServerApp,
 	mountBlobsApp,
+	mountCloudAuth,
+	mountInferenceApp,
+	mountRoomsApp,
+	mountSessionApp,
 	personal,
-	type ResolveUser,
+	recordRoomAccessOnDb,
+	requireBearerUser,
 	requireCookieOrBearerUser,
-	startBunServer,
+	type ResolveUser,
+	resolveRequestOAuthUser,
+	ServerBindings,
 } from '@epicenter/server/bun';
 import { type } from 'arktype';
+import pg from 'pg';
 import { buildEpicenterTrustedOrigins } from './worker/trusted-origins.js';
+
+/**
+ * The apps/api Bun env contract: the portable {@link ServerBindings} plus this
+ * host's process config (`DATABASE_URL`, port, origin, data dir) and the
+ * cloud-only secrets the hosted star re-requires.
+ *
+ * `ServerBindings` makes the Google credentials and `BETTER_AUTH_SECRET` optional
+ * for the single-partition instance's sake (register-when-present, ADR-0071/0073);
+ * the hub re-requires them here (its one sign-in method, plus the Better Auth
+ * signing secret it composes), so a misconfiguration fails closed at boot instead
+ * of as a downstream surprise. Unlike the Cloudflare edge (whose bindings are
+ * deploy-gated and `wrangler types`-typed), `process.env` is unchecked, so boot is
+ * the place to validate it.
+ */
+const ApiBunBindings = ServerBindings.merge({
+	DATABASE_URL: 'string',
+	'PORT?': 'string',
+	'API_PUBLIC_ORIGIN?': 'string',
+	'DATA_DIR?': 'string',
+	GOOGLE_CLIENT_ID: 'string',
+	GOOGLE_CLIENT_SECRET: 'string',
+	BETTER_AUTH_SECRET: 'string',
+});
 
 /**
  * Boot the apps/api Bun server, optionally with an injected user resolver.
@@ -55,43 +99,68 @@ import { buildEpicenterTrustedOrigins } from './worker/trusted-origins.js';
 export function startBunApiServer(
 	opts: { resolveUser?: ResolveUser } = {},
 ): void {
-	// Validate this Bun host's environment once, at boot: the library's portable
-	// secrets (`BunHostBindings` extends `ServerBindings`) and this host's own
-	// config, so a misconfiguration gets ONE descriptive error naming every
-	// missing or malformed var instead of a downstream surprise. The validated
-	// result IS the typed env handed to the Hono app: no `as`-cast over
-	// `process.env`, no lie (ADR-0066). Unlike the Cloudflare edge (whose bindings
-	// are deploy-gated and `wrangler types`-typed), `process.env` is unchecked, so
-	// boot is the place to validate it.
-	//
-	// The hosted star re-requires the secrets `ServerBindings` made optional for the
-	// instance's sake (register-when-present, ADR-0071/0073): the Google OAuth
-	// credentials (its one sign-in method) and `BETTER_AUTH_SECRET` (it composes
-	// Better Auth, the instance does not). Loosening them in the shared contract is
-	// what lets a no-auth instance boot; the hub still fails closed if any is unset.
-	const env = BunHostBindings.merge({
-		GOOGLE_CLIENT_ID: 'string',
-		GOOGLE_CLIENT_SECRET: 'string',
-		BETTER_AUTH_SECRET: 'string',
-	})(process.env);
+	// Validate this Bun host's environment once, at boot. The validated result IS
+	// the typed env handed to the Hono app: no `as`-cast over `process.env`, no
+	// lie (ADR-0066). A misconfiguration gets ONE descriptive error naming every
+	// missing or malformed var.
+	const env = ApiBunBindings(process.env);
 	if (env instanceof type.errors) {
 		console.error(`Invalid environment for the Bun server:\n${env.summary}`);
 		process.exit(1);
 	}
 
-	// apps/api partitions per user (`personal()`) and adds the content-addressed
-	// blob store; everything else is the shared Bun bootstrap.
-	const { origin, dataDir } = startBunServer({
-		env,
-		defaultPort: 8788,
-		product: 'hub',
-		ownership: personal(),
-		resolveTrustedOrigins: buildEpicenterTrustedOrigins,
-		mountExtras: (app, ownership) =>
-			mountBlobsApp(app, { ownership, auth: requireCookieOrBearerUser }),
-		// Undefined in production; `server.dev.ts` passes a dev bearer resolver.
-		resolveUser: opts.resolveUser,
+	const port = Number(env.PORT ?? 8788);
+	// The auth origin must match where the process actually listens (cookies, the
+	// OAuth issuer, the token audience all derive from it). Default to localhost
+	// on the chosen port; an operator overrides it with their domain.
+	const origin = env.API_PUBLIC_ORIGIN ?? `http://localhost:${port}`;
+
+	// One room directory of `bun:sqlite` files for this host.
+	const dataDir = resolve(env.DATA_DIR ?? './.data/rooms');
+	mkdirSync(dataDir, { recursive: true });
+	const bunRooms = createBunRooms({ dir: dataDir });
+
+	// One pool for the process; drizzle checks a client out per query and returns
+	// it, so `bun()`'s `db.connect` hands back the shared handle with a no-op close.
+	const pool = new pg.Pool({ connectionString: env.DATABASE_URL });
+	const db = createDb(pool);
+
+	const ownership = personal();
+	const app = createServerApp({
+		runtime: bun({ db, rooms: bunRooms.rooms }),
+		identity: {
+			resolveOrigin: () => origin,
+			resolveTrustedOrigins: buildEpicenterTrustedOrigins,
+		},
+		// The dev entry passes a dev bearer resolver for the parity smoke; production
+		// keeps the real OAuth bearer resolver.
+		resolveUser: opts.resolveUser ?? resolveRequestOAuthUser,
 	});
+
+	app.get('/', (c) =>
+		c.json({ product: 'hub', version: '0.1.0', runtime: 'bun' }),
+	);
+	// The cloud's relational-auth layer (Better Auth on `c.var.auth` + the auth
+	// surface), mounted before the owner-scoped surfaces read it. Host-only cookies
+	// on the Bun dev host (no cross-subdomain domain like the Worker's `.epicenter.so`).
+	mountCloudAuth(app);
+	mountSessionApp(app, { ownership, auth: requireCookieOrBearerUser });
+	mountRoomsApp(app, { ownership, recordAccess: recordRoomAccessOnDb });
+	mountInferenceApp(app, { auth: requireBearerUser, ownership });
+	mountBlobsApp(app, { ownership, auth: requireCookieOrBearerUser });
+
+	const server = Bun.serve({
+		port,
+		// Bun calls `fetch(req, server)`; route everything through the Hono app with
+		// the validated env as `c.env`. WebSocket upgrades happen inside the rooms
+		// route via the bound server (see createBunRooms), after auth runs, so they
+		// are never intercepted ahead of the auth pipeline here.
+		fetch: (req) => app.fetch(req, env),
+		websocket: bunRooms.websocket,
+	});
+	// `server` only exists once `Bun.serve` returns; hand it to the room registry
+	// so `handleUpgrade` can call `server.upgrade`.
+	bunRooms.bindServer(server);
 
 	console.log(`apps/api (Bun) listening on ${origin} (rooms in ${dataDir})`);
 }
